# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Gerencia e lida com as escritas para o serviço Kea",
        "title": "DHCP - API",
        "contact": {},
        "version": "1.0"
    },
    "host": "172.23.58.10:8005",
    "basePath": "/dhcp",
    "paths": {
        "/dhcp/conf": {
            "post": {
                "description": "Realiza a configuração do DHCP e restarta o serviço Kea",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Configurar DHCP",
                "operationId": "configDhcp",
                "parameters": [
                    {
                        "description": "Configurações do DHCP",
                        "name": "ConfigDhcp",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dhcp.DataNetworkConfigurations"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dhcp.RetornoConf"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "dhcp.DataNetworkConfigurations": {
            "type": "object",
            "properties": {
                "dnsPrimario": {
                    "type": "string"
                },
                "dnsSecundario": {
                    "type": "string"
                },
                "gateway": {
                    "type": "string"
                },
                "lease": {
                    "type": "string"
                },
                "netInterface": {
                    "type": "string"
                },
                "rangeFinal": {
                    "type": "string"
                },
                "rangeInicial": {
                    "type": "string"
                },
                "subNet": {
                    "type": "string"
                }
            }
        },
        "dhcp.RetornoConf": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "string"
                }
            }
        },
...
    }
}